Cleaner, dedicated test for parameter string writing.
