+++ tools/python/symfunc_paramgen/tests/test_outputs.py
@@ -40,77 +40,8 @@ def test_element_combinations(symfunc_type, target_combinations):
     assert myGen.element_combinations == target_combinations
 
 
-def isnotcomment(line):
-    if line[0] == '#':
-        return False
-    return True
-
-
-def test_output(tmpdir):
-    '''To detect changes in parameter generation and writing, compared to older versions.
-
-    '''
-    elems = ['S', 'Cu']
-
-    outfile_path = os.path.join(tmpdir, 'outfile.txt')
-
-    myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=6.)
-    myGen.symfunc_type = 'radial'
-    myGen.generate_radial_params(rule='imbalzano2018', mode='shift',
-                                 nb_param_pairs=5)
-    myGen.write_settings_overview(outfile_path)
-    myGen.write_parameter_strings(outfile_path)
-
-    myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=6.)
-    myGen.symfunc_type = 'radial'
-    myGen.generate_radial_params(rule='gastegger2018', mode='shift',
-                                 nb_param_pairs=5, r_lower=1.5, r_upper=5.5)
-    myGen.write_settings_overview(outfile_path)
-    myGen.write_parameter_strings(outfile_path)
-
-    myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=6.)
-    myGen.symfunc_type = 'angular_narrow'
-    myGen.generate_radial_params(rule='gastegger2018', mode='shift',
-                                 nb_param_pairs=9, r_lower=1.5)
-    myGen.zetas = [1.0, 6.0]
-    myGen.write_settings_overview(outfile_path)
-    myGen.write_parameter_strings(outfile_path)
-
-    myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=6.)
-    myGen.symfunc_type = 'angular_wide'
-    myGen.generate_radial_params(rule='gastegger2018', mode='center',
-                                 nb_param_pairs=3, r_lower=1.5)
-    myGen.zetas = [1.0, 6.0]
-    myGen.write_settings_overview(outfile_path)
-    myGen.write_parameter_strings(outfile_path)
-
-    myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=5.)
-    myGen.symfunc_type = 'weighted_radial'
-    myGen.generate_radial_params(rule='imbalzano2018', mode='center',
-                                 nb_param_pairs=5)
-    myGen.write_settings_overview(outfile_path)
-    myGen.write_parameter_strings(outfile_path)
-
-    myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=6.)
-    myGen.symfunc_type = 'weighted_angular'
-    myGen.generate_radial_params(rule='gastegger2018', mode='shift',
-                                 nb_param_pairs=3, r_lower=1.5)
-    myGen.zetas = [1.0, 6.0]
-    myGen.write_settings_overview(outfile_path)
-    myGen.write_parameter_strings(outfile_path)
-
-    # ignore comment lines, so the test does not immediately fail when only
-    # the parameter information is changed (which I would like to keep just
-    # for keeping track of the settings used in the reference output file, but
-    # is otherwise not essential)
-    with open(outfile_path) as f_out, open('reference_outputs.txt') as f_reference:
-        f_reference = filter(isnotcomment, f_reference)
-        f_out = filter(isnotcomment, f_out)
-        assert all(x == y for x, y in zip(f_out, f_reference))
-
-
 def test_setter_symfunc_type():
-    """Test if error when trying to set invalid symmetry function type
+    """Test if error when trying to set invalid symmetry function type.
     """
     elems = ['S', 'Cu']
     myGen = sfpg.SymFuncParamGenerator(elements=elems, r_cutoff=6.)
